move normalization layers.
